Moved NdHistogram to separate file
