+++ cpp/pmft/PMFTXY2D.cc
@@ -47,8 +47,8 @@ PMFTXY2D::PMFTXY2D(float x_max, float y_max, unsigned int n_x, unsigned int n_y)
     m_y_array = precomputeAxisBinCenter(m_n_y, m_dy, m_y_max);
 
     // create and populate the pcf_array
-    m_pcf_array = returnEmptyArray<float>(m_n_x * m_n_y);
-    m_bin_counts = returnEmptyArray<unsigned int>(m_n_x * m_n_y);
+    m_pcf_array = util::returnEmptyArray<float>(m_n_x * m_n_y);
+    m_bin_counts = util::returnEmptyArray<unsigned int>(m_n_x * m_n_y);
 
     // Set r_cut
     m_r_cut = sqrtf(m_x_max * m_x_max + m_y_max * m_y_max);
